修改bug

+++ src/main/java/com/xiliulou/electricity/controller/user/JsonUserJoinShareActivityRecordController.java
@@ -38,7 +38,7 @@ public class JsonUserJoinShareActivityRecordController {
      *
      */
     @PostMapping(value = "/user/joinShareActivityRecord/joinActivity")
-    public R joinActivity(@RequestParam(value = "activityId") Integer activityId,@RequestParam(value = "uid") Integer uid) {
+    public R joinActivity(@RequestParam(value = "activityId") Integer activityId,@RequestParam(value = "uid") Long uid) {
         return joinShareActivityRecordService.joinActivity(activityId,uid);
     }
