+++ src/main/java/com/xiliulou/electricity/service/JoinShareActivityRecordService.java
@@ -38,5 +38,5 @@ public interface JoinShareActivityRecordService {
 
 	R decryptScene(String scene);
 
-    R joinActivity(Integer activityId, Integer uid);
+    R joinActivity(Integer activityId, Long uid);
 }
